warnings in sdo, moxy, dbws

+++ sdo/org.eclipse.persistence.sdo/src/test/java/org/eclipse/persistence/testing/sdo/helper/xmlhelper/loadandsave/withoutxsd/LoadAndSaveNestedSchemaTypeTestCases.java
@@ -60,7 +60,7 @@ public class LoadAndSaveNestedSchemaTypeTestCases extends LoadAndSaveUnknownTest
         assertNotNull(prop);
         assertEquals(SDOConstants.SDO_INT, prop.getType());
         assertTrue(value instanceof Integer);
-        assertEquals(Integer.valueOf(10), ((Integer)value));
+        assertEquals(Integer.valueOf(10), value);
 
     }
 }
